not cache peer verification logic
