+++ package.json
@@ -5,7 +5,7 @@
   "author": "Ryo Ota <nwtgck@nwtgck.org> (https://github.com/nwtgck)",
   "scripts": {
     "serve": "npm run generate-files && vue-cli-service serve",
-    "build": "npm run generate-files && vue-cli-service build",
+    "build": "npm run generate-files && vue-cli-service build && test -f ./dist/js/piping-ui-auth*.js",
     "test:unit": "vue-cli-service test:unit",
     "lint": "vue-cli-service lint",
     "build-sw": "webpack --config src/service-worker/webpack.config.js src/service-worker/src/sw.ts",
